Refactors search result actions
