+++ test/actions/helpers.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import { formatQuery } from '../../src/actions/helpers.js';
+import { formatTags } from '../../src/actions/helpers.js';
 import { expect } from 'chai';
 
 describe('Helpers', () => {
@@ -10,7 +10,7 @@ describe('Helpers', () => {
     {id: 'amenity:pool', displayName: 'pool'}
   ];
   it('formatQuery: returns an object with geography and amenity keys', (done) => {
-    const res = formatQuery(tags);
+    const res = formatTags(tags);
     expect(Object.keys(res)).to.deep.equal(['geography', 'amenity']);
     expect(res.geography).to.deep.equal(['geo:geonames:12345']);
     expect(res.amenity).to.deep.equal(['amenity:wifi', 'amenity:pool']);
@@ -20,7 +20,7 @@ describe('Helpers', () => {
     const tags = [
       {id: 'geo:geonames:12345', displayName: 'spain'}
     ];
-    const res = formatQuery(tags);
+    const res = formatTags(tags);
     expect(Object.keys(res)).to.deep.equal(['geography']);
     expect(res.geography).to.deep.equal(['geo:geonames:12345']);
     done();
