+++ test/actions/search-results.test.js
@@ -1,9 +1,6 @@
 import {
   BUSY_SEARCHING,
-  SAVE_SEARCH_RESULT_ID,
-  RECEIVE_SEARCH_RESULT
-  // TILES_ADD_TILES,
-  // SEARCH_ERROR
+  SAVE_SEARCH_RESULT_ID
 } from '../../src/constants/actionTypes';
 import moment from 'moment';
 
@@ -88,109 +85,6 @@ describe('actions', function () {
           done();
         })
         .catch(done);
-    });
-  });
-  describe('fetchQuerySearchResults', function () {
-    afterEach(function (done) {
-      simple.restore();
-      done();
-    });
-    const json = {
-      data: {
-        viewer: {
-          searchResult: {
-            items: [{type: 'packageOffer'}, {type: 'tile'}]
-          }
-        }
-      }
-    };
-
-    // Does not happen anymore
-    // it('no displayedItems -> should call the addTiles and receiveSearchResult actions with the items', function (done) {
-    //   const expectedActions = [
-    //     { type: 'TILES_ADD_TILES', tileArray: undefined },
-    //     {
-    //       type: 'RECEIVE_SEARCH_RESULT',
-    //       items: [{
-    //         type: 'packageOffer'
-    //       }],
-    //       initialSearch: true,
-    //       append: true
-    //     }
-    //   ];
-    //   var stub = simple.mock(graphqlService, 'query').resolveWith(json);
-    //   const store = mockStore(initialState);
-    //   store.dispatch(actions.fetchQuerySearchResults('1', 1, 2, 0))
-    //     .then(() => {
-    //       expect(store.getActions()).to.deep.equal(expectedActions);
-    //       expect(stub.callCount).to.equal(1);
-    //       done();
-    //     })
-    //     .catch(done);
-    // });
-    it('existing displayedItems -> should call only the receiveSearchResult action with the items', function (done) {
-      const expectedActions = [
-        {
-          type: RECEIVE_SEARCH_RESULT,
-          items: [{type: 'packageOffer'}, {type: 'tile'}],
-          initialSearch: false,
-          append: false
-        }
-      ];
-      var stub = simple.mock(graphqlService, 'query').resolveWith(json);
-      const initialStateWithResults = {
-        ...initialState,
-        search: {
-          ...initialState.search,
-          displayedItems: [{}]
-        }
-      };
-      const store = mockStore(initialStateWithResults);
-      store.dispatch(actions.fetchQuerySearchResults('1', 1, 2, 0))
-        .then(() => {
-          expect(store.getActions()).to.deep.equal(expectedActions);
-          expect(stub.callCount).to.equal(1);
-          done();
-        })
-        .catch(done);
-    });
-    // We are currently not displaying an error anymore
-    // it('attempt > 15 -> should dispatch a search error', function (done) {
-    //   simple.mock(graphqlService, 'query').resolveWith(json);
-    //   const expectedActions = [
-    //     {
-    //       type: SEARCH_ERROR,
-    //       error: 'Something went wrong and no results were found'
-    //     }
-    //   ];
-    //   const store = mockStore(initialState);
-    //   store.dispatch(actions.fetchQuerySearchResults('1', 1, 2, 16))
-    //     .then(() => {
-    //       console.log('actions', store.getActions());
-    //       expect(store.getActions()).to.deep.equal(expectedActions);
-    //       done();
-    //     })
-    //     .catch(done);
-    // });
-    it('no items returned from graphql -> should poll for more results', function (done) {
-      this.timeout(10100);
-      const noItems = {data: { viewer: { searchResult: { items: [] } } }};
-      simple.mock(graphqlService, 'query')
-        .resolveWith(noItems);
-      const store = mockStore(initialState);
-      const stub = simple.mock(store, 'dispatch').callOriginal();
-      store.dispatch(actions.fetchQuerySearchResults('1', 1, 2, 1));
-
-      graphqlService.query.lastCall.returned
-        .then(json => {
-          expect(json).to.equal(noItems);
-          setTimeout(function () {
-            expect(stub.lastCall.arg.name).to.equal('fetchQuerySearchResults_anonymousFn');
-            expect(graphqlService.query.callCount).to.equal(10); // exits after the 10th attempt (starts at 1)
-            done();
-          }, 10000);
-        })
-       .catch(done);
     });
   });
 });
